+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Allow treating a missing highlight language as error
 - Handle more editors with change detection in `zola serve`
 - Add argument to `zola serve` to write HTML files to disk
+- Add optional parsing of Markdown definition lists
 
 
 ## 0.19.2 (2024-08-15)